HW1/requirements.txt
ultralytics==8.2.0
torch==2.4.0
torchvision==0.19.0
numpy==2.0.1
pandas==2.2.2
pillow==10.3.0
pyyaml==6.0.1
